// src/types.ts
export interface mClothingType {
    hat: string | null;
    hair: string | null;
    top: string | null;
    bottom: string | null;
    footwear: string | null;
    handItem: string | null;
  }
  
  export interface fClothingType {
    hat: string | null;
    hair: string | null;
    outfit: string | null;
    skirt: string | null;
    footwear: string | null;
    handItem: string | null;
  }
  